api_gateway/requirements.txt
uvicorn==0.34.0
fastapi==0.115.1
httpx==0.28.1
pyjwt==2.10.1
pydantic-settings==2.0.2